DEMOSYS-861 - schedule job always deleting source files

+++ Model/Queue/Consumer.php
@@ -140,12 +140,14 @@ class Consumer
         $dataPack->setReload(1);
         $this->process->loadFiles($dataPack);
         //delete source files if it's an uploaded package
-        if ($this->fileSystem->isExists($dataPack->getDataPackLocation())) {
-            $this->fileSystem->deleteDirectory($dataPack->getDataPackLocation());
-            //delete the archive that is from a mac compress process
-            $macFile = $this->fileSystem->getParentDirectory($dataPack->getDataPackLocation())."/__MACOSX";
-            if ($this->fileSystem->isExists($macFile)) {
-                $this->fileSystem->deleteDirectory($macFile);
+        if($dataPack->deleteSourceFiles()){
+            if ($this->fileSystem->isExists($dataPack->getDataPackLocation())) {
+                $this->fileSystem->deleteDirectory($dataPack->getDataPackLocation());
+                //delete the archive that is from a mac compress process
+                $macFile = $this->fileSystem->getParentDirectory($dataPack->getDataPackLocation())."/__MACOSX";
+                if ($this->fileSystem->isExists($macFile)) {
+                    $this->fileSystem->deleteDirectory($macFile);
+                }
             }
         }
     }
@@ -170,6 +172,7 @@ class Consumer
         $dataPack->setIsDefaultWebsite($data['isDefaultWebsite']);
         $dataPack->setHost($data['host']);
         $dataPack->setJobId($data['jobid']);
+        $dataPack->setDeleteSourceFiles($data['deleteSourceFiles']);
         return $dataPack;
     }
 }
